refactor(nav): move helper types to mod
